Build fix
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@26672 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/PDMAsyncCompletionFileNormal.cpp
@@ -711,7 +711,7 @@ static int pdmacFileAioMgrNormalProcessTaskList(PPDMACTASKFILE pTaskHead,
     {
         PPDMACTASKFILE pCurr = pTaskHead;
 
-        if (!pdmacFileBwMgrIsTransferAllowed(pEndpoint->pBwMgr, pCurr->DataSeg.cbSeg))
+        if (!pdmacFileBwMgrIsTransferAllowed(pEndpoint->pBwMgr, (uint32_t)pCurr->DataSeg.cbSeg))
         {
             pAioMgr->fBwLimitReached = true;
             break;
